+++ basics/account-data/steel/Cargo.toml
@@ -1,3 +1,9 @@
 [workspace]
 members = ["program"]
 resolver = "2"
+
+[workspace.dependencies]
+solana-program = "1.18.17"
+steel = "2.0"
+bytemuck = "1.14"
+num_enum = "0.7"
