+++ basics/account-data/steel/api/src/lib.rs
@@ -1,0 +1,10 @@
+mod instruction;
+mod state;
+
+use steel::*;
+
+declare_id!("z7msBPQHDJjTvdQRoEcKyENgXDhSRYeHieN1ZMTqo35");
+
+pub mod prelude {
+    pub use crate::state::*;
+}
